+++ cardano-crypto-class/src/Cardano/Crypto/KES/Simple.hs
@@ -213,7 +213,7 @@ instance
       convertSK =
         fmap (fromMaybe (error "unsoundPureSignKeyKESToSoundSignKeyKES: deserialisation failed"))
           . rawDeserialiseSignKeyDSIGNM
-          . rawSerialiseSignKeyDSIGN
+          . rawEncodeFixedSized
 
 instance
   (UnsoundDSIGNMAlgorithm d, KnownNat t, KESAlgorithm (SimpleKES d t)) =>
@@ -228,7 +228,7 @@ instance
 
   rawDeserialiseSignKeyKESWith allocator bs
     | let duration = fromIntegral @Natural @Int (natVal (Proxy :: Proxy t))
-          sizeKey = fromIntegral @Word @Int (signKeySizeDSIGN (Proxy :: Proxy d))
+          sizeKey = fromIntegral @Word @Int (fixedSize (Proxy :: Proxy (SignKeyDSIGN d)))
     , skbs <- splitsAt (replicate duration sizeKey) bs
     , length skbs == duration =
         runMaybeT $ do
@@ -250,7 +250,7 @@ instance
   rawDecodeFixedSized bs = do
     guardFixedSized (Proxy @(VerKeyKES (SimpleKES d t))) bs
     let duration = fromIntegral @Natural @Int (natVal (Proxy :: Proxy t))
-        sizeKey = fromIntegral @Word @Int (verKeySizeDSIGN (Proxy :: Proxy d))
+        sizeKey = fromIntegral @Word @Int (fixedSize (Proxy :: Proxy (VerKeyDSIGN d)))
         vkbs = splitsAt (replicate duration sizeKey) bs
     unless (length vkbs == duration) $
       fail "VerKeyKES (SimpleKES d t): vkbs length is not equal to duration"
@@ -280,7 +280,7 @@ instance
   rawDecodeFixedSized bs = do
     guardFixedSized (Proxy @(UnsoundPureSignKeyKES (SimpleKES d t))) bs
     let duration = fromIntegral @Natural @Int (natVal (Proxy :: Proxy t))
-        sizeKey = fromIntegral @Word @Int (signKeySizeDSIGN (Proxy :: Proxy d))
+        sizeKey = fromIntegral @Word @Int (fixedSize (Proxy :: Proxy (SignKeyDSIGN d)))
         skbs = splitsAt (replicate duration sizeKey) bs
     unless (length skbs == duration) $
       fail "UnsoundPureSignKeyKES (SimpleKES d t): vkbs length is not equal to duration"